+++ test/extensions/transport_sockets/tls/integration/BUILD
@@ -1,0 +1,59 @@
+licenses(["notice"])  # Apache 2
+
+load(
+    "//bazel:envoy_build_system.bzl",
+    "envoy_cc_fuzz_test",
+    "envoy_cc_test",
+    "envoy_cc_test_library",
+    "envoy_package",
+    "envoy_proto_library",
+    "envoy_select_hot_restart",
+    "envoy_sh_test",
+)
+load(
+    "//source/extensions:all_extensions.bzl",
+    "envoy_all_extensions",
+)
+
+envoy_package()
+
+envoy_cc_test_library(
+    name = "ssl_integration_lib",
+    hdrs = [
+        "ssl_integration_test.h",
+    ],
+    deps = [
+        "//source/common/http:header_map_lib",
+        "//source/extensions/filters/listener/tls_inspector:config",
+        "//source/extensions/transport_sockets/tls:config",
+        "//test/integration:http_integration_lib",
+        "//test/mocks/server:server_mocks",
+        "//test/test_common:utility_lib",
+    ],
+)
+
+envoy_cc_test(
+    name = "ssl_integration_test",
+    srcs = [
+        "ssl_integration_test.cc",
+        "ssl_integration_test.h",
+    ],
+    data = [
+        "//test/config/integration/certs",
+    ],
+    deps = [
+        "//source/common/event:dispatcher_includes",
+        "//source/common/event:dispatcher_lib",
+        "//source/common/network:connection_lib",
+        "//source/common/network:utility_lib",
+        "//source/extensions/filters/listener/tls_inspector:config",
+        "//source/extensions/transport_sockets/tls:config",
+        "//source/extensions/transport_sockets/tls:context_config_lib",
+        "//source/extensions/transport_sockets/tls:context_lib",
+        "//test/integration:http_integration_lib",
+        "//test/mocks/secret:secret_mocks",
+        "//test/test_common:utility_lib",
+        "@envoy_api//envoy/config/transport_socket/tap/v2alpha:tap_cc",
+        "@envoy_api//envoy/data/tap/v2alpha:transport_cc",
+    ],
+)
