+++ test/integration/BUILD
@@ -535,32 +535,6 @@ envoy_cc_test(
 )
 
 envoy_cc_test(
-    name = "ssl_integration_test",
-    srcs = [
-        "ssl_integration_test.cc",
-        "ssl_integration_test.h",
-    ],
-    data = [
-        "//test/config/integration/certs",
-    ],
-    deps = [
-        ":http_integration_lib",
-        "//source/common/event:dispatcher_includes",
-        "//source/common/event:dispatcher_lib",
-        "//source/common/network:connection_lib",
-        "//source/common/network:utility_lib",
-        "//source/extensions/filters/listener/tls_inspector:config",
-        "//source/extensions/transport_sockets/tls:config",
-        "//source/extensions/transport_sockets/tls:context_config_lib",
-        "//source/extensions/transport_sockets/tls:context_lib",
-        "//test/mocks/secret:secret_mocks",
-        "//test/test_common:utility_lib",
-        "@envoy_api//envoy/config/transport_socket/tap/v2alpha:tap_cc",
-        "@envoy_api//envoy/data/tap/v2alpha:transport_cc",
-    ],
-)
-
-envoy_cc_test(
     name = "tcp_proxy_integration_test",
     srcs = [
         "tcp_proxy_integration_test.cc",
@@ -639,7 +613,6 @@ envoy_cc_test(
 envoy_cc_test(
     name = "xfcc_integration_test",
     srcs = [
-        "ssl_integration_test.h",
         "xfcc_integration_test.cc",
         "xfcc_integration_test.h",
     ],
@@ -651,6 +624,7 @@ envoy_cc_test(
         "//source/common/http:header_map_lib",
         "//source/extensions/filters/listener/tls_inspector:config",
         "//source/extensions/transport_sockets/tls:config",
+        "//test/extensions/transport_sockets/tls/integration:ssl_integration_lib",
         "//test/mocks/server:server_mocks",
         "//test/test_common:utility_lib",
     ],
