+++ test/integration/xfcc_integration_test.cc
@@ -15,6 +15,7 @@
 #include "extensions/transport_sockets/tls/context_manager_impl.h"
 #include "extensions/transport_sockets/tls/ssl_socket.h"
 
+#include "test/extensions/transport_sockets/tls/integration/ssl_integration_test.h"
 #include "test/test_common/network_utility.h"
 #include "test/test_common/printers.h"
 #include "test/test_common/test_base.h"
@@ -22,7 +23,6 @@
 
 #include "gmock/gmock.h"
 #include "integration.h"
-#include "ssl_integration_test.h"
 #include "utility.h"
 
 namespace Envoy {
